LayoutTests/storage/indexeddb/modern/basic-add.html

This test does basic testing of IDBObjectStore.add(), making sure that an attempt to overwrite an already-existing key fails with the appropriate error.<br>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

var request = window.indexedDB.open("NewDatabaseAddTestDatabase");

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

request.onupgradeneeded = function(event) {
    alert("Upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);
    
    var tx = request.transaction;
    var db = event.target.result;

    alert(tx + " - " + tx.mode);
    alert(db);

    var os = db.createObjectStore("TestObjectStore");
    var putRequest1 = os.add("bar", "foo");
    var putRequest2 = os.add("baz", "foo");

    putRequest1.onsuccess = function(event) {
        alert("put 1 succeeded - key was '" + putRequest1.result + "'");
    }

    putRequest1.onerror = function(event) {
        alert("put 1 unexpectedly failed - " + event);
        done();
    }

    putRequest2.onsuccess = function(event) {
        alert("put 2 unexpectedly succeeded - key was '" + putRequest2.result + "'");
        done();
    }

    putRequest2.onerror = function(event) {
        alert("put 2 failed - " + event.type);

        var getRequest = os.get("foo");
        
        getRequest.onsuccess = function(event) {
            alert("get succeeded - key was '" + getRequest.result + "'");
        }

        getRequest.onerror = function(event) {
            alert("get unexpectedly failed - " + event.type);
            done();
        }
        
        event.stopPropagation();
        event.preventDefault();
    }
        
    tx.onabort = function(event) {
        alert("version change transaction unexpected abort");
        done();
    }

    tx.oncomplete = function(event) {
        alert("version change transaction completed");
        done();
    }

    tx.onerror = function(event) {
        alert("version change transaction unexpected error - " + event);
        done();
    }
}
</script>
